refactor(rln): split /circuit/mod.rs into multiple files
