+++ rln/src/circuit/mod.rs
@@ -4,428 +4,23 @@ pub(crate) mod error;
 pub(crate) mod iden3calc;
 pub(crate) mod qap;
 
-#[cfg(not(target_arch = "wasm32"))]
-use std::sync::LazyLock;
-use std::{ops::Deref, sync::Arc};
+mod backend;
+mod graph;
+mod id_secret;
+mod types;
+mod zkey;
 
-use ark_bn254::{
-    Bn254, Fq as ArkFq, Fq2 as ArkFq2, Fr as ArkFr, G1Affine as ArkG1Affine,
-    G1Projective as ArkG1Projective, G2Affine as ArkG2Affine, G2Projective as ArkG2Projective,
+pub use backend::ArkGroth16Backend;
+#[cfg(not(target_arch = "wasm32"))]
+pub use graph::{default_graph_multi, default_graph_single};
+pub use graph::{graph_from_raw, Graph};
+pub(crate) use graph::{CalcWitness, CalcWitnessPartial};
+pub(crate) use id_secret::FrOrSecret;
+pub use id_secret::IdSecret;
+pub use types::{
+    Curve, Fq, Fq2, Fr, G1Affine, G1Projective, G2Affine, G2Projective, PartialProof, Proof,
+    ProvingKey, VerifyingKey, Zkey, COMPRESS_PROOF_SIZE, DEFAULT_MAX_OUT, DEFAULT_TREE_DEPTH,
 };
-use ark_ff::{Field, PrimeField, UniformRand};
-use ark_groth16::{
-    prepare_verifying_key, PreparedVerifyingKey as ArkPreparedVerifyingKey, Proof as ArkProof,
-    ProvingKey as ArkProvingKey, VerifyingKey as ArkVerifyingKey,
-};
-use ark_relations::r1cs::ConstraintMatrices;
-use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
-use rand::Rng;
-use ruint::aliases::U256;
-use zeroize::{Zeroize, ZeroizeOnDrop};
-
-use self::error::{GraphReadError, WitnessCalcError, ZKeyReadError};
-use crate::{
-    circuit::iden3calc::{graph::Node, storage::deserialize_witnesscalc_graph, InputSignalsInfo},
-    partial_proof::PartialProof as ArkPartialProof,
-};
-
+pub use zkey::zkey_from_raw;
 #[cfg(not(target_arch = "wasm32"))]
-const GRAPH_BYTES_SINGLE: &[u8] = include_bytes!("../../resources/tree_depth_20/graph.bin");
-
-#[cfg(not(target_arch = "wasm32"))]
-const ARKZKEY_BYTES_SINGLE: &[u8] =
-    include_bytes!("../../resources/tree_depth_20/rln_final.arkzkey");
-
-#[cfg(not(target_arch = "wasm32"))]
-const GRAPH_BYTES_MULTI: &[u8] =
-    include_bytes!("../../resources/tree_depth_20/multi_message_id/max_out_4/graph.bin");
-
-#[cfg(not(target_arch = "wasm32"))]
-const ARKZKEY_BYTES_MULTI: &[u8] =
-    include_bytes!("../../resources/tree_depth_20/multi_message_id/max_out_4/rln_final.arkzkey");
-
-#[cfg(not(target_arch = "wasm32"))]
-static ARKZKEY_SINGLE: LazyLock<Arc<Zkey>> = LazyLock::new(|| {
-    Arc::new(
-        read_arkzkey_from_bytes_uncompressed(ARKZKEY_BYTES_SINGLE)
-            .expect("Default Single zkey must be valid"),
-    )
-});
-
-#[cfg(not(target_arch = "wasm32"))]
-static ARKZKEY_MULTI: LazyLock<Arc<Zkey>> = LazyLock::new(|| {
-    Arc::new(
-        read_arkzkey_from_bytes_uncompressed(ARKZKEY_BYTES_MULTI)
-            .expect("Default Multi zkey must be valid"),
-    )
-});
-
-#[cfg(not(target_arch = "wasm32"))]
-static GRAPH_SINGLE: LazyLock<Arc<Graph>> = LazyLock::new(|| {
-    Arc::new(
-        graph_from_raw(GRAPH_BYTES_SINGLE, Some(DEFAULT_TREE_DEPTH), None)
-            .expect("Default Single graph must be valid"),
-    )
-});
-
-#[cfg(not(target_arch = "wasm32"))]
-static GRAPH_MULTI: LazyLock<Arc<Graph>> = LazyLock::new(|| {
-    Arc::new(
-        graph_from_raw(
-            GRAPH_BYTES_MULTI,
-            Some(DEFAULT_TREE_DEPTH),
-            Some(DEFAULT_MAX_OUT),
-        )
-        .expect("Default Multi graph must be valid"),
-    )
-});
-
-pub const DEFAULT_MAX_OUT: usize = 4;
-pub const DEFAULT_TREE_DEPTH: usize = 20;
-pub const COMPRESS_PROOF_SIZE: usize = 128;
-
-// The following types define the pairing friendly elliptic curve, the underlying finite fields and groups default to this module
-// Note that proofs are serialized assuming Fr to be 4x8 = 32 bytes in size. Hence, changing to a curve with different encoding will make proof verification to fail
-
-/// BN254 pairing-friendly elliptic curve.
-pub type Curve = Bn254;
-
-/// Scalar field Fr of the BN254 curve.
-pub type Fr = ArkFr;
-
-/// Base field Fq of the BN254 curve.
-pub type Fq = ArkFq;
-
-/// Quadratic extension field element for the BN254 curve.
-pub type Fq2 = ArkFq2;
-
-/// Affine representation of a G1 group element on the BN254 curve.
-pub type G1Affine = ArkG1Affine;
-
-/// Projective representation of a G1 group element on the BN254 curve.
-pub type G1Projective = ArkG1Projective;
-
-/// Affine representation of a G2 group element on the BN254 curve.
-pub type G2Affine = ArkG2Affine;
-
-/// Projective representation of a G2 group element on the BN254 curve.
-pub type G2Projective = ArkG2Projective;
-
-/// Groth16 proof for the BN254 curve.
-pub type Proof = ArkProof<Curve>;
-
-/// Partial Groth16 proof for the BN254 curve.
-pub type PartialProof = ArkPartialProof<Curve>;
-
-/// Proving key for the Groth16 proof system.
-pub type ProvingKey = ArkProvingKey<Curve>;
-
-/// Combining the proving key and constraint matrices.
-pub type Zkey = (ArkProvingKey<Curve>, ConstraintMatrices<Fr>);
-
-/// Verifying key for the Groth16 proof system.
-pub type VerifyingKey = ArkVerifyingKey<Curve>;
-
-/// Witness calculator graph.
-///
-/// Contains the deserialized computation graph used for witness calculation.
-/// Parsing this once and reusing it avoids repeated deserialization overhead.
-#[derive(Clone, Debug)]
-pub struct Graph {
-    pub(crate) nodes: Vec<Node>,
-    pub(crate) signals: Vec<usize>,
-    pub(crate) input_mapping: InputSignalsInfo,
-    pub(crate) tree_depth: usize,
-    pub(crate) max_out: usize,
-}
-
-/// Loads the zkey from raw bytes
-pub fn zkey_from_raw(zkey_data: &[u8]) -> Result<Zkey, ZKeyReadError> {
-    if zkey_data.is_empty() {
-        return Err(ZKeyReadError::EmptyBytes);
-    }
-
-    let proving_key_and_matrices = read_arkzkey_from_bytes_uncompressed(zkey_data)?;
-
-    Ok(proving_key_and_matrices)
-}
-
-/// Parses the witness calculator graph from raw bytes
-pub fn graph_from_raw(
-    graph_data: &[u8],
-    expected_tree_depth: Option<usize>,
-    expected_max_out: Option<usize>,
-) -> Result<Graph, GraphReadError> {
-    if graph_data.is_empty() {
-        return Err(GraphReadError::EmptyBytes);
-    }
-
-    let (nodes, signals, input_mapping) =
-        deserialize_witnesscalc_graph(std::io::Cursor::new(graph_data))?;
-
-    let tree_depth = {
-        let (_, depth) = input_mapping
-            .get("pathElements")
-            .ok_or_else(|| GraphReadError::MissingSignal("pathElements".to_string()))?;
-
-        if let Some(expected) = expected_tree_depth {
-            if expected != *depth {
-                return Err(GraphReadError::TreeDepthMismatch {
-                    expected,
-                    actual: *depth,
-                });
-            }
-        }
-
-        *depth
-    };
-
-    let max_out = {
-        let (_, count) = input_mapping
-            .get("messageId")
-            .ok_or_else(|| GraphReadError::MissingSignal("messageId".to_string()))?;
-
-        if let Some(expected) = expected_max_out {
-            if expected != *count {
-                return Err(GraphReadError::MaxOutMismatch {
-                    expected,
-                    actual: *count,
-                });
-            }
-        }
-
-        *count
-    };
-
-    Ok(Graph {
-        nodes,
-        signals,
-        input_mapping,
-        tree_depth,
-        max_out,
-    })
-}
-
-/// Loads default Single zkey
-#[cfg(not(target_arch = "wasm32"))]
-pub fn default_zkey_single() -> &'static Arc<Zkey> {
-    &ARKZKEY_SINGLE
-}
-
-/// Loads default Multi zkey
-#[cfg(not(target_arch = "wasm32"))]
-pub fn default_zkey_multi() -> &'static Arc<Zkey> {
-    &ARKZKEY_MULTI
-}
-
-/// Loads default Single graph
-#[cfg(not(target_arch = "wasm32"))]
-pub fn default_graph_single() -> &'static Arc<Graph> {
-    &GRAPH_SINGLE
-}
-
-/// Loads default Multi graph
-#[cfg(not(target_arch = "wasm32"))]
-pub fn default_graph_multi() -> &'static Arc<Graph> {
-    &GRAPH_MULTI
-}
-
-/// Secret field-element wrapper zeroized on drop.
-#[derive(
-    Debug, Zeroize, ZeroizeOnDrop, Clone, PartialEq, CanonicalSerialize, CanonicalDeserialize,
-)]
-pub struct IdSecret(Fr);
-
-impl IdSecret {
-    pub fn rand<R: Rng + ?Sized>(rng: &mut R) -> Self {
-        let mut fr = Fr::rand(rng);
-        let res = Self::from(&mut fr);
-        // No need to zeroize fr (already zeroiz'ed in from implementation)
-        #[allow(clippy::let_and_return)]
-        res
-    }
-
-    /// Warning: this can leak the secret value
-    /// Warning: Leaked value is of type 'U256' which implement Copy (every copy will not be zeroized)
-    pub(crate) fn to_u256(&self) -> U256 {
-        let mut big_int = self.0.into_bigint();
-        let res = U256::from_limbs(big_int.0);
-        big_int.zeroize(); // wipe the secret limbs after copying them into the leaked U256
-        res
-    }
-}
-
-impl From<&mut Fr> for IdSecret {
-    fn from(value: &mut Fr) -> Self {
-        let id_secret = Self(*value);
-        value.zeroize(); // clear the caller-owned source Fr after the secret moved into IdSecret
-        id_secret
-    }
-}
-
-impl Deref for IdSecret {
-    type Target = Fr;
-
-    /// Deref to &Fr
-    ///
-    /// Warning: this can leak the secret value
-    /// Warning: Leaked value is of type 'Fr' which implement Copy (every copy will not be zeroized)
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-#[derive(Debug, Clone, Zeroize, ZeroizeOnDrop)]
-pub(crate) enum FrOrSecret {
-    IdSecret(IdSecret),
-    Fr(Fr),
-}
-
-impl From<Fr> for FrOrSecret {
-    fn from(value: Fr) -> Self {
-        FrOrSecret::Fr(value)
-    }
-}
-
-impl From<IdSecret> for FrOrSecret {
-    fn from(value: IdSecret) -> Self {
-        FrOrSecret::IdSecret(value)
-    }
-}
-
-// The following functions and structs are based on code from ark-zkey:
-// https://github.com/zkmopro/ark-zkey/blob/main/src/lib.rs#L106
-
-#[derive(CanonicalSerialize, CanonicalDeserialize, Clone, Debug, PartialEq)]
-struct SerializableProvingKey(ArkProvingKey<Curve>);
-
-#[derive(CanonicalSerialize, CanonicalDeserialize, Clone, Debug, PartialEq)]
-struct SerializableConstraintMatrices<F: Field> {
-    num_instance_variables: usize,
-    num_witness_variables: usize,
-    num_constraints: usize,
-    a_num_non_zero: usize,
-    b_num_non_zero: usize,
-    c_num_non_zero: usize,
-    a: SerializableMatrix<F>,
-    b: SerializableMatrix<F>,
-    c: SerializableMatrix<F>,
-}
-
-#[derive(CanonicalSerialize, CanonicalDeserialize, Clone, Debug, PartialEq)]
-struct SerializableMatrix<F: Field> {
-    pub data: Vec<Vec<(F, usize)>>,
-}
-
-fn read_arkzkey_from_bytes_uncompressed(arkzkey_data: &[u8]) -> Result<Zkey, ZKeyReadError> {
-    if arkzkey_data.is_empty() {
-        return Err(ZKeyReadError::EmptyBytes);
-    }
-
-    let mut cursor = std::io::Cursor::new(arkzkey_data);
-
-    let serialized_proving_key =
-        SerializableProvingKey::deserialize_uncompressed_unchecked(&mut cursor)?;
-
-    let serialized_constraint_matrices =
-        SerializableConstraintMatrices::deserialize_uncompressed_unchecked(&mut cursor)?;
-
-    let proving_key: ProvingKey = serialized_proving_key.0;
-    let constraint_matrices: ConstraintMatrices<Fr> = ConstraintMatrices {
-        num_instance_variables: serialized_constraint_matrices.num_instance_variables,
-        num_witness_variables: serialized_constraint_matrices.num_witness_variables,
-        num_constraints: serialized_constraint_matrices.num_constraints,
-        a_num_non_zero: serialized_constraint_matrices.a_num_non_zero,
-        b_num_non_zero: serialized_constraint_matrices.b_num_non_zero,
-        c_num_non_zero: serialized_constraint_matrices.c_num_non_zero,
-        a: serialized_constraint_matrices.a.data,
-        b: serialized_constraint_matrices.b.data,
-        c: serialized_constraint_matrices.c.data,
-    };
-    let zkey = (proving_key, constraint_matrices);
-
-    Ok(zkey)
-}
-
-#[derive(Clone, Debug)]
-pub struct ArkGroth16Backend {
-    pub(crate) zkey: Arc<Zkey>,
-    pub(crate) graph: Arc<Graph>,
-    pub(crate) pvk: ArkPreparedVerifyingKey<Curve>,
-}
-
-impl ArkGroth16Backend {
-    pub fn new(zkey: impl Into<Arc<Zkey>>, graph: impl Into<Arc<Graph>>) -> Self {
-        let zkey = zkey.into();
-        let graph = graph.into();
-        let pvk = prepare_verifying_key(&zkey.0.vk);
-        Self { zkey, graph, pvk }
-    }
-}
-
-/// Witness type can calculate a full circuit witness directly from its fields.
-pub(crate) trait CalcWitness {
-    fn calc_witness(&self, graph: &Graph) -> Result<Vec<Fr>, WitnessCalcError>;
-}
-
-/// Witness type can calculate a partial circuit witness (unknown dynamic fields become `None`).
-pub(crate) trait CalcWitnessPartial {
-    fn calc_witness_partial(&self, graph: &Graph) -> Result<Vec<Option<Fr>>, WitnessCalcError>;
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn test_empty_zkey_and_graph() {
-        let err = zkey_from_raw(&[]).unwrap_err();
-        assert!(matches!(err, ZKeyReadError::EmptyBytes));
-
-        let err = graph_from_raw(&[], None, None).err().unwrap();
-        assert!(matches!(err, GraphReadError::EmptyBytes));
-
-        let err = read_arkzkey_from_bytes_uncompressed(&[]).unwrap_err();
-        assert!(matches!(err, ZKeyReadError::EmptyBytes));
-    }
-
-    #[test]
-    fn test_tree_depth_mismatch() {
-        let err = graph_from_raw(GRAPH_BYTES_SINGLE, Some(DEFAULT_TREE_DEPTH + 1), None)
-            .err()
-            .unwrap();
-        assert!(matches!(err, GraphReadError::TreeDepthMismatch { .. }));
-    }
-
-    #[test]
-    fn test_max_out_mismatch() {
-        let err = graph_from_raw(
-            GRAPH_BYTES_MULTI,
-            Some(DEFAULT_TREE_DEPTH),
-            Some(DEFAULT_MAX_OUT + 1),
-        )
-        .err()
-        .unwrap();
-        assert!(matches!(err, GraphReadError::MaxOutMismatch { .. }));
-    }
-
-    #[test]
-    fn test_missing_signal_rejected() {
-        let single = graph_from_raw(GRAPH_BYTES_SINGLE, None, None).unwrap();
-        assert!(single.input_mapping.contains_key("pathElements"));
-        assert!(single.input_mapping.contains_key("messageId"));
-
-        let multi = graph_from_raw(GRAPH_BYTES_MULTI, None, None).unwrap();
-        assert!(multi.input_mapping.contains_key("pathElements"));
-        assert!(multi.input_mapping.contains_key("messageId"));
-    }
-
-    #[test]
-    fn test_id_secret_from_fr_zeroizes_source() {
-        let mut fr = Fr::from(42);
-        let id_secret = IdSecret::from(&mut fr);
-
-        assert_ne!(fr, Fr::from(42));
-        assert_eq!(*id_secret, Fr::from(42));
-    }
-}
+pub use zkey::{default_zkey_multi, default_zkey_single};
